candidate-recommendation-engine/requirements.txt
streamlit
python-dotenv
pandas
numpy
openai
sentence-transformers
pdfplumber
python-docx
scikit-learn
